+++ libraries/AP_NavEKF2/AP_NavEKF2_ControlState.cpp
@@ -18,32 +18,6 @@
 
 extern const AP_HAL::HAL& hal;
 
-// Check basic filter health metrics and return a consolidated health status
-bool NavEKF2_core::healthy(void) const
-{
-    uint8_t faultInt;
-    getFilterFaults(faultInt);
-    if (faultInt > 0) {
-        return false;
-    }
-    if (velTestRatio > 1 && posTestRatio > 1 && hgtTestRatio > 1) {
-        // all three metrics being above 1 means the filter is
-        // extremely unhealthy.
-        return false;
-    }
-    // Give the filter a second to settle before use
-    if ((imuSampleTime_ms - ekfStartTime_ms) < 1000 ) {
-        return false;
-    }
-    // barometer and position innovations must be within limits when on-ground
-    float horizErrSq = sq(innovVelPos[3]) + sq(innovVelPos[4]);
-    if (onGround && (fabsf(innovVelPos[5]) > 1.0f || horizErrSq > 1.0f)) {
-        return false;
-    }
-
-    // all OK
-    return true;
-}
 
 // Control filter mode transitions
 void NavEKF2_core::controlFilterModes()
@@ -248,175 +222,5 @@ bool NavEKF2_core::assume_zero_sideslip(void) const
     // traction is lost
     return _ahrs->get_fly_forward() && _ahrs->get_vehicle_class() != AHRS_VEHICLE_GROUND;
 }
-
-// return the innovations for the NED Pos, NED Vel, XYZ Mag and Vtas measurements
-void  NavEKF2_core::getInnovations(Vector3f &velInnov, Vector3f &posInnov, Vector3f &magInnov, float &tasInnov, float &yawInnov) const
-{
-    velInnov.x = innovVelPos[0];
-    velInnov.y = innovVelPos[1];
-    velInnov.z = innovVelPos[2];
-    posInnov.x = innovVelPos[3];
-    posInnov.y = innovVelPos[4];
-    posInnov.z = innovVelPos[5];
-    magInnov.x = 1e3f*innovMag[0]; // Convert back to sensor units
-    magInnov.y = 1e3f*innovMag[1]; // Convert back to sensor units
-    magInnov.z = 1e3f*innovMag[2]; // Convert back to sensor units
-    tasInnov   = innovVtas;
-    yawInnov   = innovYaw;
-}
-
-// return the innovation consistency test ratios for the velocity, position, magnetometer and true airspeed measurements
-// this indicates the amount of margin available when tuning the various error traps
-// also return the current offsets applied to the GPS position measurements
-void  NavEKF2_core::getVariances(float &velVar, float &posVar, float &hgtVar, Vector3f &magVar, float &tasVar, Vector2f &offset) const
-{
-    velVar   = sqrtf(velTestRatio);
-    posVar   = sqrtf(posTestRatio);
-    hgtVar   = sqrtf(hgtTestRatio);
-    magVar.x = sqrtf(magTestRatio.x);
-    magVar.y = sqrtf(magTestRatio.y);
-    magVar.z = sqrtf(magTestRatio.z);
-    tasVar   = sqrtf(tasTestRatio);
-    offset   = gpsPosGlitchOffsetNE;
-}
-
-
-/*
-return the filter fault status as a bitmasked integer
- 0 = quaternions are NaN
- 1 = velocities are NaN
- 2 = badly conditioned X magnetometer fusion
- 3 = badly conditioned Y magnetometer fusion
- 5 = badly conditioned Z magnetometer fusion
- 6 = badly conditioned airspeed fusion
- 7 = badly conditioned synthetic sideslip fusion
- 7 = filter is not initialised
-*/
-void  NavEKF2_core::getFilterFaults(uint8_t &faults) const
-{
-    faults = (stateStruct.quat.is_nan()<<0 |
-              stateStruct.velocity.is_nan()<<1 |
-              faultStatus.bad_xmag<<2 |
-              faultStatus.bad_ymag<<3 |
-              faultStatus.bad_zmag<<4 |
-              faultStatus.bad_airspeed<<5 |
-              faultStatus.bad_sideslip<<6 |
-              !statesInitialised<<7);
-}
-
-/*
-return filter timeout status as a bitmasked integer
- 0 = position measurement timeout
- 1 = velocity measurement timeout
- 2 = height measurement timeout
- 3 = magnetometer measurement timeout
- 4 = true airspeed measurement timeout
- 5 = unassigned
- 6 = unassigned
- 7 = unassigned
-*/
-void  NavEKF2_core::getFilterTimeouts(uint8_t &timeouts) const
-{
-    timeouts = (posTimeout<<0 |
-                velTimeout<<1 |
-                hgtTimeout<<2 |
-                magTimeout<<3 |
-                tasTimeout<<4);
-}
-
-/*
-return filter function status as a bitmasked integer
- 0 = attitude estimate valid
- 1 = horizontal velocity estimate valid
- 2 = vertical velocity estimate valid
- 3 = relative horizontal position estimate valid
- 4 = absolute horizontal position estimate valid
- 5 = vertical position estimate valid
- 6 = terrain height estimate valid
- 7 = constant position mode
-*/
-void  NavEKF2_core::getFilterStatus(nav_filter_status &status) const
-{
-    // init return value
-    status.value = 0;
-
-    bool doingFlowNav = (PV_AidingMode == AID_RELATIVE) && flowDataValid;
-    bool doingWindRelNav = !tasTimeout && assume_zero_sideslip();
-    bool doingNormalGpsNav = !posTimeout && (PV_AidingMode == AID_ABSOLUTE);
-    bool notDeadReckoning = (PV_AidingMode == AID_ABSOLUTE);
-    bool someVertRefData = (!velTimeout && useGpsVertVel) || !hgtTimeout;
-    bool someHorizRefData = !(velTimeout && posTimeout && tasTimeout) || doingFlowNav;
-    bool optFlowNavPossible = flowDataValid && (frontend._fusionModeGPS == 3);
-    bool gpsNavPossible = !gpsNotAvailable && (frontend._fusionModeGPS <= 2);
-    bool filterHealthy = healthy() && tiltAlignComplete && yawAlignComplete;
-
-    // set individual flags
-    status.flags.attitude = !stateStruct.quat.is_nan() && filterHealthy;   // attitude valid (we need a better check)
-    status.flags.horiz_vel = someHorizRefData && notDeadReckoning && filterHealthy;      // horizontal velocity estimate valid
-    status.flags.vert_vel = someVertRefData && filterHealthy;        // vertical velocity estimate valid
-    status.flags.horiz_pos_rel = ((doingFlowNav && gndOffsetValid) || doingWindRelNav || doingNormalGpsNav) && notDeadReckoning && filterHealthy;   // relative horizontal position estimate valid
-    status.flags.horiz_pos_abs = doingNormalGpsNav && notDeadReckoning && filterHealthy; // absolute horizontal position estimate valid
-    status.flags.vert_pos = !hgtTimeout && filterHealthy;            // vertical position estimate valid
-    status.flags.terrain_alt = gndOffsetValid && filterHealthy;		// terrain height estimate valid
-    status.flags.const_pos_mode = (PV_AidingMode == AID_NONE) && filterHealthy;     // constant position mode
-    status.flags.pred_horiz_pos_rel = (optFlowNavPossible || gpsNavPossible) && filterHealthy; // we should be able to estimate a relative position when we enter flight mode
-    status.flags.pred_horiz_pos_abs = gpsNavPossible && filterHealthy; // we should be able to estimate an absolute position when we enter flight mode
-    status.flags.takeoff_detected = takeOffDetected; // takeoff for optical flow navigation has been detected
-    status.flags.takeoff = expectGndEffectTakeoff; // The EKF has been told to expect takeoff and is in a ground effect mitigation mode
-    status.flags.touchdown = expectGndEffectTouchdown; // The EKF has been told to detect touchdown and is in a ground effect mitigation mode
-    status.flags.using_gps = (imuSampleTime_ms - lastPosPassTime_ms) < 4000;
-}
-
-// send an EKF_STATUS message to GCS
-void NavEKF2_core::send_status_report(mavlink_channel_t chan)
-{
-    // get filter status
-    nav_filter_status filt_state;
-    getFilterStatus(filt_state);
-
-    // prepare flags
-    uint16_t flags = 0;
-    if (filt_state.flags.attitude) {
-        flags |= EKF_ATTITUDE;
-    }
-    if (filt_state.flags.horiz_vel) {
-        flags |= EKF_VELOCITY_HORIZ;
-    }
-    if (filt_state.flags.vert_vel) {
-        flags |= EKF_VELOCITY_VERT;
-    }
-    if (filt_state.flags.horiz_pos_rel) {
-        flags |= EKF_POS_HORIZ_REL;
-    }
-    if (filt_state.flags.horiz_pos_abs) {
-        flags |= EKF_POS_HORIZ_ABS;
-    }
-    if (filt_state.flags.vert_pos) {
-        flags |= EKF_POS_VERT_ABS;
-    }
-    if (filt_state.flags.terrain_alt) {
-        flags |= EKF_POS_VERT_AGL;
-    }
-    if (filt_state.flags.const_pos_mode) {
-        flags |= EKF_CONST_POS_MODE;
-    }
-    if (filt_state.flags.pred_horiz_pos_rel) {
-        flags |= EKF_PRED_POS_HORIZ_REL;
-    }
-    if (filt_state.flags.pred_horiz_pos_abs) {
-        flags |= EKF_PRED_POS_HORIZ_ABS;
-    }
-
-    // get variances
-    float velVar, posVar, hgtVar, tasVar;
-    Vector3f magVar;
-    Vector2f offset;
-    getVariances(velVar, posVar, hgtVar, magVar, tasVar, offset);
-
-    // send message
-    mavlink_msg_ekf_status_report_send(chan, flags, velVar, posVar, hgtVar, magVar.length(), tasVar);
-
-}
-
 
 #endif // HAL_CPU_CLASS
