+++ libraries/AP_NavEKF2/AP_NavEKF2_Mag.cpp
@@ -105,115 +105,6 @@ void NavEKF2_core::alignYawGPS()
 }
 
 /********************************************************
-*            GET STATES/PARAMS FUNCTIONS                *
-********************************************************/
-
-// return earth magnetic field estimates in measurement units / 1000
-void NavEKF2_core::getMagNED(Vector3f &magNED) const
-{
-    magNED = stateStruct.earth_magfield * 1000.0f;
-}
-
-// return body magnetic field estimates in measurement units / 1000
-void NavEKF2_core::getMagXYZ(Vector3f &magXYZ) const
-{
-    magXYZ = stateStruct.body_magfield*1000.0f;
-}
-/********************************************************
-*            SET STATES/PARAMS FUNCTIONS                *
-********************************************************/
-
-
-/********************************************************
-*                      READ SENSORS                     *
-********************************************************/
-// return magnetometer offsets
-// return true if offsets are valid
-bool NavEKF2_core::getMagOffsets(Vector3f &magOffsets) const
-{
-    // compass offsets are valid if we have finalised magnetic field initialisation and magnetic field learning is not prohibited and primary compass is valid
-    if (secondMagYawInit && (frontend._magCal != 2) && _ahrs->get_compass()->healthy(0)) {
-        magOffsets = _ahrs->get_compass()->get_offsets(0) - stateStruct.body_magfield*1000.0f;
-        return true;
-    } else {
-        magOffsets = _ahrs->get_compass()->get_offsets(0);
-        return false;
-    }
-}
-
-// check for new magnetometer data and update store measurements if available
-void NavEKF2_core::readMagData()
-{
-    if (use_compass() && _ahrs->get_compass()->last_update_usec() != lastMagUpdate_ms) {
-        // store time of last measurement update
-        lastMagUpdate_ms = _ahrs->get_compass()->last_update_usec();
-
-        // estimate of time magnetometer measurement was taken, allowing for delays
-        magMeasTime_ms = imuSampleTime_ms - frontend.magDelay_ms;
-
-        // read compass data and scale to improve numerical conditioning
-        magDataNew.mag = _ahrs->get_compass()->get_field() * 0.001f;
-
-        // check for consistent data between magnetometers
-        consistentMagData = _ahrs->get_compass()->consistent();
-
-        // check if compass offsets have been changed and adjust EKF bias states to maintain consistent innovations
-        if (_ahrs->get_compass()->healthy(0)) {
-            Vector3f nowMagOffsets = _ahrs->get_compass()->get_offsets(0);
-            bool changeDetected = (!is_equal(nowMagOffsets.x,lastMagOffsets.x) || !is_equal(nowMagOffsets.y,lastMagOffsets.y) || !is_equal(nowMagOffsets.z,lastMagOffsets.z));
-            // Ignore bias changes before final mag field and yaw initialisation, as there may have been a compass calibration
-            if (changeDetected && secondMagYawInit) {
-                stateStruct.body_magfield.x += (nowMagOffsets.x - lastMagOffsets.x) * 0.001f;
-                stateStruct.body_magfield.y += (nowMagOffsets.y - lastMagOffsets.y) * 0.001f;
-                stateStruct.body_magfield.z += (nowMagOffsets.z - lastMagOffsets.z) * 0.001f;
-            }
-            lastMagOffsets = nowMagOffsets;
-        }
-
-        // save magnetometer measurement to buffer to be fused later
-        magDataNew.time_ms = magMeasTime_ms;
-        StoreMag();
-    }
-}
-// store magnetometer data in a history array
-void NavEKF2_core::StoreMag()
-{
-    if (magStoreIndex >= OBS_BUFFER_LENGTH) {
-        magStoreIndex = 0;
-    }
-    storedMag[magStoreIndex] = magDataNew;
-    magStoreIndex += 1;
-}
-
-// return newest un-used magnetometer data that has fallen behind the fusion time horizon
-// if no un-used data is available behind the fusion horizon, return false
-bool NavEKF2_core::RecallMag()
-{
-    mag_elements dataTemp;
-    mag_elements dataTempZero;
-    dataTempZero.time_ms = 0;
-    uint32_t temp_ms = 0;
-    for (uint8_t i=0; i<OBS_BUFFER_LENGTH; i++) {
-        dataTemp = storedMag[i];
-        // find a measurement older than the fusion time horizon that we haven't checked before
-        if (dataTemp.time_ms != 0 && dataTemp.time_ms <= imuDataDelayed.time_ms) {
-            // zero the time stamp so we won't use it again
-            storedMag[i]=dataTempZero;
-            // Find the most recent non-stale measurement that meets the time horizon criteria
-            if (((imuDataDelayed.time_ms - dataTemp.time_ms) < 500) && dataTemp.time_ms > temp_ms) {
-                magDataDelayed = dataTemp;
-                temp_ms = dataTemp.time_ms;
-            }
-        }
-    }
-    if (temp_ms != 0) {
-        return true;
-    } else {
-        return false;
-    }
-}
-
-/********************************************************
 *                   FUSE MEASURED_DATA                  *
 ********************************************************/
