+++ libraries/AP_NavEKF2/AP_NavEKF2_OptFlow.cpp
@@ -18,177 +18,6 @@
 
 extern const AP_HAL::HAL& hal;
 
-// return data for debugging optical flow fusion
-void NavEKF2_core::getFlowDebug(float &varFlow, float &gndOffset, float &flowInnovX, float &flowInnovY, float &auxInnov, float &HAGL, float &rngInnov, float &range, float &gndOffsetErr) const
-{
-    varFlow = max(flowTestRatio[0],flowTestRatio[1]);
-    gndOffset = terrainState;
-    flowInnovX = innovOptFlow[0];
-    flowInnovY = innovOptFlow[1];
-    auxInnov = auxFlowObsInnov;
-    HAGL = terrainState - stateStruct.position.z;
-    rngInnov = innovRng;
-    range = rngMea;
-    gndOffsetErr = sqrtf(Popt); // note Popt is constrained to be non-negative in EstimateTerrainOffset()
-}
-
-// provides the height limit to be observed by the control loops
-// returns false if no height limiting is required
-// this is needed to ensure the vehicle does not fly too high when using optical flow navigation
-bool NavEKF2_core::getHeightControlLimit(float &height) const
-{
-    // only ask for limiting if we are doing optical flow navigation
-    if (frontend._fusionModeGPS == 3) {
-        // If are doing optical flow nav, ensure the height above ground is within range finder limits after accounting for vehicle tilt and control errors
-        height = max(float(_rng.max_distance_cm()) * 0.007f - 1.0f, 1.0f);
-        return true;
-    } else {
-        return false;
-    }
-}
-
-
-// Read the range finder and take new measurements if available
-// Read at 20Hz and apply a median filter
-void NavEKF2_core::readRangeFinder(void)
-{
-    uint8_t midIndex;
-    uint8_t maxIndex;
-    uint8_t minIndex;
-    // get theoretical correct range when the vehicle is on the ground
-    rngOnGnd = _rng.ground_clearance_cm() * 0.01f;
-    if (_rng.status() == RangeFinder::RangeFinder_Good && (imuSampleTime_ms - lastRngMeasTime_ms) > 50) {
-        // store samples and sample time into a ring buffer
-        rngMeasIndex ++;
-        if (rngMeasIndex > 2) {
-            rngMeasIndex = 0;
-        }
-        storedRngMeasTime_ms[rngMeasIndex] = imuSampleTime_ms;
-        storedRngMeas[rngMeasIndex] = _rng.distance_cm() * 0.01f;
-        // check for three fresh samples and take median
-        bool sampleFresh[3];
-        for (uint8_t index = 0; index <= 2; index++) {
-            sampleFresh[index] = (imuSampleTime_ms - storedRngMeasTime_ms[index]) < 500;
-        }
-        if (sampleFresh[0] && sampleFresh[1] && sampleFresh[2]) {
-            if (storedRngMeas[0] > storedRngMeas[1]) {
-                minIndex = 1;
-                maxIndex = 0;
-            } else {
-                maxIndex = 0;
-                minIndex = 1;
-            }
-            if (storedRngMeas[2] > storedRngMeas[maxIndex]) {
-                midIndex = maxIndex;
-            } else if (storedRngMeas[2] < storedRngMeas[minIndex]) {
-                midIndex = minIndex;
-            } else {
-                midIndex = 2;
-            }
-            rngMea = max(storedRngMeas[midIndex],rngOnGnd);
-            newDataRng = true;
-            rngValidMeaTime_ms = imuSampleTime_ms;
-        } else if (onGround) {
-            // if on ground and no return, we assume on ground range
-            rngMea = rngOnGnd;
-            newDataRng = true;
-            rngValidMeaTime_ms = imuSampleTime_ms;
-        } else {
-            newDataRng = false;
-        }
-        lastRngMeasTime_ms =  imuSampleTime_ms;
-    }
-}
-
-// write the raw optical flow measurements
-// this needs to be called externally.
-void NavEKF2_core::writeOptFlowMeas(uint8_t &rawFlowQuality, Vector2f &rawFlowRates, Vector2f &rawGyroRates, uint32_t &msecFlowMeas)
-{
-    // The raw measurements need to be optical flow rates in radians/second averaged across the time since the last update
-    // The PX4Flow sensor outputs flow rates with the following axis and sign conventions:
-    // A positive X rate is produced by a positive sensor rotation about the X axis
-    // A positive Y rate is produced by a positive sensor rotation about the Y axis
-    // This filter uses a different definition of optical flow rates to the sensor with a positive optical flow rate produced by a
-    // negative rotation about that axis. For example a positive rotation of the flight vehicle about its X (roll) axis would produce a negative X flow rate
-    flowMeaTime_ms = imuSampleTime_ms;
-    // calculate bias errors on flow sensor gyro rates, but protect against spikes in data
-    // reset the accumulated body delta angle and time
-    // don't do the calculation if not enough time lapsed for a reliable body rate measurement
-    if (delTimeOF > 0.01f) {
-        flowGyroBias.x = 0.99f * flowGyroBias.x + 0.01f * constrain_float((rawGyroRates.x - delAngBodyOF.x/delTimeOF),-0.1f,0.1f);
-        flowGyroBias.y = 0.99f * flowGyroBias.y + 0.01f * constrain_float((rawGyroRates.y - delAngBodyOF.y/delTimeOF),-0.1f,0.1f);
-        delAngBodyOF.zero();
-        delTimeOF = 0.0f;
-    }
-    // check for takeoff if relying on optical flow and zero measurements until takeoff detected
-    // if we haven't taken off - constrain position and velocity states
-    if (frontend._fusionModeGPS == 3) {
-        detectOptFlowTakeoff();
-    }
-    // calculate rotation matrices at mid sample time for flow observations
-    stateStruct.quat.rotation_matrix(Tbn_flow);
-    Tnb_flow = Tbn_flow.transposed();
-    // don't use data with a low quality indicator or extreme rates (helps catch corrupt sensor data)
-    if ((rawFlowQuality > 0) && rawFlowRates.length() < 4.2f && rawGyroRates.length() < 4.2f) {
-        // correct flow sensor rates for bias
-        omegaAcrossFlowTime.x = rawGyroRates.x - flowGyroBias.x;
-        omegaAcrossFlowTime.y = rawGyroRates.y - flowGyroBias.y;
-        // write uncorrected flow rate measurements that will be used by the focal length scale factor estimator
-        // note correction for different axis and sign conventions used by the px4flow sensor
-        ofDataNew.flowRadXY = - rawFlowRates; // raw (non motion compensated) optical flow angular rate about the X axis (rad/sec)
-        // write flow rate measurements corrected for body rates
-        ofDataNew.flowRadXYcomp.x = ofDataNew.flowRadXY.x + omegaAcrossFlowTime.x;
-        ofDataNew.flowRadXYcomp.y = ofDataNew.flowRadXY.y + omegaAcrossFlowTime.y;
-        // record time last observation was received so we can detect loss of data elsewhere
-        flowValidMeaTime_ms = imuSampleTime_ms;
-        // estimate sample time of the measurement
-        ofDataNew.time_ms = imuSampleTime_ms - frontend._flowDelay_ms - frontend.flowTimeDeltaAvg_ms/2;
-        // Save data to buffer
-        StoreOF();
-        // Check for data at the fusion time horizon
-        newDataFlow = RecallOF();
-    }
-}
-
-// store OF data in a history array
-void NavEKF2_core::StoreOF()
-{
-    if (ofStoreIndex >= OBS_BUFFER_LENGTH) {
-        ofStoreIndex = 0;
-    }
-    storedOF[ofStoreIndex] = ofDataNew;
-    ofStoreIndex += 1;
-}
-
-// return newest un-used optical flow data that has fallen behind the fusion time horizon
-// if no un-used data is available behind the fusion horizon, return false
-bool NavEKF2_core::RecallOF()
-{
-    of_elements dataTemp;
-    of_elements dataTempZero;
-    dataTempZero.time_ms = 0;
-    uint32_t temp_ms = 0;
-    for (uint8_t i=0; i<OBS_BUFFER_LENGTH; i++) {
-        dataTemp = storedOF[i];
-        // find a measurement older than the fusion time horizon that we haven't checked before
-        if (dataTemp.time_ms != 0 && dataTemp.time_ms <= imuDataDelayed.time_ms) {
-            // zero the time stamp so we won't use it again
-            storedOF[i]=dataTempZero;
-            // Find the most recent non-stale measurement that meets the time horizon criteria
-            if (((imuDataDelayed.time_ms - dataTemp.time_ms) < 500) && dataTemp.time_ms > temp_ms) {
-                ofDataDelayed = dataTemp;
-                temp_ms = dataTemp.time_ms;
-            }
-        }
-    }
-    if (temp_ms != 0) {
-        return true;
-    } else {
-        return false;
-    }
-}
-
-
 // select fusion of optical flow measurements
 void NavEKF2_core::SelectFlowFusion()
 {
