+++ libraries/AP_NavEKF2/AP_NavEKF2_PosVelNED.cpp
@@ -108,70 +108,8 @@ bool NavEKF2_core::resetHeightDatum(void)
 
 
 /********************************************************
-*            GET STATES/PARAMS FUNCTIONS                *
+*                 GET PARAMS FUNCTIONS                  *
 ********************************************************/
-
-// return NED velocity in m/s
-//
-void NavEKF2_core::getVelNED(Vector3f &vel) const
-{
-    vel = outputDataNew.velocity;
-}
-
-// This returns the specific forces in the NED frame
-void NavEKF2_core::getAccelNED(Vector3f &accelNED) const {
-    accelNED = velDotNED;
-    accelNED.z -= GRAVITY_MSS;
-}
-
-// return the Z-accel bias estimate in m/s^2
-void NavEKF2_core::getAccelZBias(float &zbias) const {
-    if (dtIMUavg > 0) {
-        zbias = stateStruct.accel_zbias / dtIMUavg;
-    } else {
-        zbias = 0;
-    }
-}
-
-// Return the last calculated NED position relative to the reference point (m).
-// if a calculated solution is not available, use the best available data and return false
-bool NavEKF2_core::getPosNED(Vector3f &pos) const
-{
-    // The EKF always has a height estimate regardless of mode of operation
-    pos.z = outputDataNew.position.z;
-    // There are three modes of operation, absolute position (GPS fusion), relative position (optical flow fusion) and constant position (no position estimate available)
-    nav_filter_status status;
-    getFilterStatus(status);
-    if (status.flags.horiz_pos_abs || status.flags.horiz_pos_rel) {
-        // This is the normal mode of operation where we can use the EKF position states
-        pos.x = outputDataNew.position.x;
-        pos.y = outputDataNew.position.y;
-        return true;
-    } else {
-        // In constant position mode the EKF position states are at the origin, so we cannot use them as a position estimate
-        if(validOrigin) {
-            if ((_ahrs->get_gps().status() >= AP_GPS::GPS_OK_FIX_2D)) {
-                // If the origin has been set and we have GPS, then return the GPS position relative to the origin
-                const struct Location &gpsloc = _ahrs->get_gps().location();
-                Vector2f tempPosNE = location_diff(EKF_origin, gpsloc);
-                pos.x = tempPosNE.x;
-                pos.y = tempPosNE.y;
-                return false;
-            } else {
-                // If no GPS fix is available, all we can do is provide the last known position
-                pos.x = outputDataNew.position.x;
-                pos.y = outputDataNew.position.y;
-                return false;
-            }
-        } else {
-            // If the origin has not been set, then we have no means of providing a relative position
-            pos.x = 0.0f;
-            pos.y = 0.0f;
-            return false;
-        }
-    }
-    return false;
-}
 
 // return the horizontal speed limit in m/s set by optical flow sensor limits
 // return the scale factor to be applied to navigation velocity gains to compensate for increase in velocity noise with height when using optical flow
@@ -188,53 +126,6 @@ void NavEKF2_core::getEkfControlLimits(float &ekfGndSpdLimit, float &ekfNavVelGa
     }
 }
 
-// Return the last calculated latitude, longitude and height in WGS-84
-// If a calculated location isn't available, return a raw GPS measurement
-// The status will return true if a calculation or raw measurement is available
-// The getFilterStatus() function provides a more detailed description of data health and must be checked if data is to be used for flight control
-bool NavEKF2_core::getLLH(struct Location &loc) const
-{
-    if(validOrigin) {
-        // Altitude returned is an absolute altitude relative to the WGS-84 spherioid
-        loc.alt = EKF_origin.alt - outputDataNew.position.z*100;
-        loc.flags.relative_alt = 0;
-        loc.flags.terrain_alt = 0;
-
-        // there are three modes of operation, absolute position (GPS fusion), relative position (optical flow fusion) and constant position (no aiding)
-        nav_filter_status status;
-        getFilterStatus(status);
-        if (status.flags.horiz_pos_abs || status.flags.horiz_pos_rel) {
-            loc.lat = EKF_origin.lat;
-            loc.lng = EKF_origin.lng;
-            location_offset(loc, outputDataNew.position.x, outputDataNew.position.y);
-            return true;
-        } else {
-            // we could be in constant position mode  becasue the vehicle has taken off without GPS, or has lost GPS
-            // in this mode we cannot use the EKF states to estimate position so will return the best available data
-            if ((_ahrs->get_gps().status() >= AP_GPS::GPS_OK_FIX_2D)) {
-                // we have a GPS position fix to return
-                const struct Location &gpsloc = _ahrs->get_gps().location();
-                loc.lat = gpsloc.lat;
-                loc.lng = gpsloc.lng;
-                return true;
-            } else {
-                // if no GPS fix, provide last known position before entering the mode
-                location_offset(loc, lastKnownPositionNE.x, lastKnownPositionNE.y);
-                return false;
-            }
-        }
-    } else {
-        // If no origin has been defined for the EKF, then we cannot use its position states so return a raw
-        // GPS reading if available and return false
-        if ((_ahrs->get_gps().status() >= AP_GPS::GPS_OK_FIX_3D)) {
-            const struct Location &gpsloc = _ahrs->get_gps().location();
-            loc = gpsloc;
-            loc.flags.relative_alt = 0;
-            loc.flags.terrain_alt = 0;
-        }
-        return false;
-    }
-}
 
 // return the LLH location of the filters NED origin
 bool NavEKF2_core::getOriginLLH(struct Location &loc) const
@@ -243,14 +134,6 @@ bool NavEKF2_core::getOriginLLH(struct Location &loc) const
         loc = EKF_origin;
     }
     return validOrigin;
-}
-
-// return the estimated height above ground level
-bool NavEKF2_core::getHAGL(float &HAGL) const
-{
-    HAGL = terrainState - outputDataNew.position.z;
-    // If we know the terrain offset and altitude, then we have a valid height above ground estimate
-    return !hgtTimeout && gndOffsetValid && healthy();
 }
 
 /********************************************************
@@ -296,305 +179,6 @@ uint8_t NavEKF2_core::setInhibitGPS(void)
         return 2;
     } else {
         return 1;
-    }
-}
-
-
-/********************************************************
-*                      READ SENSORS                     *
-********************************************************/
-
-bool NavEKF2_core::readDeltaVelocity(uint8_t ins_index, Vector3f &dVel, float &dVel_dt) {
-    const AP_InertialSensor &ins = _ahrs->get_ins();
-
-    if (ins_index < ins.get_accel_count()) {
-        ins.get_delta_velocity(ins_index,dVel);
-        dVel_dt = max(ins.get_delta_velocity_dt(ins_index),1.0e-4f);
-        return true;
-    }
-    return false;
-}
-
-// check for new valid GPS data and update stored measurement if available
-void NavEKF2_core::readGpsData()
-{
-    // check for new GPS data
-    if ((_ahrs->get_gps().last_message_time_ms() != lastTimeGpsReceived_ms) &&
-            (_ahrs->get_gps().status() >= AP_GPS::GPS_OK_FIX_3D))
-    {
-        // store fix time from previous read
-        secondLastGpsTime_ms = lastTimeGpsReceived_ms;
-
-        // get current fix time
-        lastTimeGpsReceived_ms = _ahrs->get_gps().last_message_time_ms();
-
-        // estimate when the GPS fix was valid, allowing for GPS processing and other delays
-        // ideally we should be using a timing signal from the GPS receiver to set this time
-        gpsDataNew.time_ms = lastTimeGpsReceived_ms - frontend._gpsDelay_ms;
-
-        // read the NED velocity from the GPS
-        gpsDataNew.vel = _ahrs->get_gps().velocity();
-
-        // Use the speed accuracy from the GPS if available, otherwise set it to zero.
-        // Apply a decaying envelope filter with a 5 second time constant to the raw speed accuracy data
-        float alpha = constrain_float(0.0002f * (lastTimeGpsReceived_ms - secondLastGpsTime_ms),0.0f,1.0f);
-        gpsSpdAccuracy *= (1.0f - alpha);
-        float gpsSpdAccRaw;
-        if (!_ahrs->get_gps().speed_accuracy(gpsSpdAccRaw)) {
-            gpsSpdAccuracy = 0.0f;
-        } else {
-            gpsSpdAccuracy = max(gpsSpdAccuracy,gpsSpdAccRaw);
-        }
-
-        // check if we have enough GPS satellites and increase the gps noise scaler if we don't
-        if (_ahrs->get_gps().num_sats() >= 6 && (PV_AidingMode == AID_ABSOLUTE)) {
-            gpsNoiseScaler = 1.0f;
-        } else if (_ahrs->get_gps().num_sats() == 5 && (PV_AidingMode == AID_ABSOLUTE)) {
-            gpsNoiseScaler = 1.4f;
-        } else { // <= 4 satellites or in constant position mode
-            gpsNoiseScaler = 2.0f;
-        }
-
-        // Check if GPS can output vertical velocity and set GPS fusion mode accordingly
-        if (_ahrs->get_gps().have_vertical_velocity() && frontend._fusionModeGPS == 0) {
-            useGpsVertVel = true;
-        } else {
-            useGpsVertVel = false;
-        }
-
-        // Monitor quality of the GPS velocity data for alignment
-        if (PV_AidingMode != AID_ABSOLUTE) {
-            gpsQualGood = calcGpsGoodToAlign();
-        }
-
-        // read latitutde and longitude from GPS and convert to local NE position relative to the stored origin
-        // If we don't have an origin, then set it to the current GPS coordinates
-        const struct Location &gpsloc = _ahrs->get_gps().location();
-        if (validOrigin) {
-            gpsDataNew.pos = location_diff(EKF_origin, gpsloc);
-        } else if (gpsQualGood) {
-            // Set the NE origin to the current GPS position
-            setOrigin();
-            // Now we know the location we have an estimate for the magnetic field declination and adjust the earth field accordingly
-            alignMagStateDeclination();
-            // Set the height of the NED origin to ‘height of baro height datum relative to GPS height datum'
-            EKF_origin.alt = gpsloc.alt - baroDataNew.hgt;
-            // We are by definition at the origin at the instant of alignment so set NE position to zero
-            gpsDataNew.pos.zero();
-            // If GPS useage isn't explicitly prohibited, we switch to absolute position mode
-            if (isAiding && frontend._fusionModeGPS != 3) {
-                PV_AidingMode = AID_ABSOLUTE;
-                // Initialise EKF position and velocity states
-                ResetPosition();
-                ResetVelocity();
-            }
-        }
-
-        // calculate a position offset which is applied to NE position and velocity wherever it is used throughout code to allow GPS position jumps to be accommodated gradually
-        decayGpsOffset();
-
-        // save measurement to buffer to be fused later
-        StoreGPS();
-
-        // declare GPS available for use
-        gpsNotAvailable = false;
-    }
-
-    // We need to handle the case where GPS is lost for a period of time that is too long to dead-reckon
-    // If that happens we need to put the filter into a constant position mode, reset the velocity states to zero
-    // and use the last estimated position as a synthetic GPS position
-
-    // check if we can use opticalflow as a backup
-    bool optFlowBackupAvailable = (flowDataValid && !hgtTimeout);
-
-    // Set GPS time-out threshold depending on whether we have an airspeed sensor to constrain drift
-    uint16_t gpsRetryTimeout_ms = useAirspeed() ? frontend.gpsRetryTimeUseTAS_ms : frontend.gpsRetryTimeNoTAS_ms;
-
-    // Set the time that copters will fly without a GPS lock before failing the GPS and switching to a non GPS mode
-    uint16_t gpsFailTimeout_ms = optFlowBackupAvailable ? frontend.gpsFailTimeWithFlow_ms : gpsRetryTimeout_ms;
-
-    // If we haven't received GPS data for a while and we are using it for aiding, then declare the position and velocity data as being timed out
-    if (imuSampleTime_ms - lastTimeGpsReceived_ms > gpsFailTimeout_ms) {
-
-        // Let other processes know that GPS i snota vailable and that a timeout has occurred
-        posTimeout = true;
-        velTimeout = true;
-        gpsNotAvailable = true;
-
-        // If we are currently reliying on GPS for navigation, then we need to switch to a non-GPS mode of operation
-        if (PV_AidingMode == AID_ABSOLUTE) {
-
-            // If we don't have airspeed or sideslip assumption or optical flow to constrain drift, then go into constant position mode.
-            // If we can do optical flow nav (valid flow data and height above ground estimate), then go into flow nav mode.
-            if (!useAirspeed() && !assume_zero_sideslip()) {
-                if (optFlowBackupAvailable) {
-                    // we can do optical flow only nav
-                    frontend._fusionModeGPS = 3;
-                    PV_AidingMode = AID_RELATIVE;
-                } else {
-                    // store the current position
-                    lastKnownPositionNE.x = stateStruct.position.x;
-                    lastKnownPositionNE.y = stateStruct.position.y;
-
-                    // put the filter into constant position mode
-                    PV_AidingMode = AID_NONE;
-
-                    // reset all glitch states
-                    gpsPosGlitchOffsetNE.zero();
-                    gpsVelGlitchOffset.zero();
-
-                    // Reset the velocity and position states
-                    ResetVelocity();
-                    ResetPosition();
-
-                    // Reset the normalised innovation to avoid false failing the bad position fusion test
-                    velTestRatio = 0.0f;
-                    posTestRatio = 0.0f;
-                }
-            }
-        }
-    }
-
-    // If not aiding we synthesise the GPS measurements at the last known position
-    if (PV_AidingMode == AID_NONE) {
-        if (imuSampleTime_ms - gpsDataNew.time_ms > 200) {
-            gpsDataNew.pos.x = lastKnownPositionNE.x;
-            gpsDataNew.pos.y = lastKnownPositionNE.y;
-            gpsDataNew.time_ms = imuSampleTime_ms-frontend._gpsDelay_ms;
-            // save measurement to buffer to be fused later
-            StoreGPS();
-        }
-    }
-
-}
-
-
-// store GPS data in a history array
-void NavEKF2_core::StoreGPS()
-{
-    if (gpsStoreIndex >= OBS_BUFFER_LENGTH) {
-        gpsStoreIndex = 0;
-    }
-    storedGPS[gpsStoreIndex] = gpsDataNew;
-    gpsStoreIndex += 1;
-}
-
-// return newest un-used GPS data that has fallen behind the fusion time horizon
-// if no un-used data is available behind the fusion horizon, return false
-bool NavEKF2_core::RecallGPS()
-{
-    gps_elements dataTemp;
-    gps_elements dataTempZero;
-    dataTempZero.time_ms = 0;
-    uint32_t temp_ms = 0;
-    for (uint8_t i=0; i<OBS_BUFFER_LENGTH; i++) {
-        dataTemp = storedGPS[i];
-        // find a measurement older than the fusion time horizon that we haven't checked before
-        if (dataTemp.time_ms != 0 && dataTemp.time_ms <= imuDataDelayed.time_ms) {
-            // zero the time stamp so we won't use it again
-            storedGPS[i]=dataTempZero;
-            // Find the most recent non-stale measurement that meets the time horizon criteria
-            if (((imuDataDelayed.time_ms - dataTemp.time_ms) < 500) && dataTemp.time_ms > temp_ms) {
-                gpsDataDelayed = dataTemp;
-                temp_ms = dataTemp.time_ms;
-            }
-        }
-    }
-    if (temp_ms != 0) {
-        return true;
-    } else {
-        return false;
-    }
-}
-
-// check for new altitude measurement data and update stored measurement if available
-void NavEKF2_core::readHgtData()
-{
-    // check to see if baro measurement has changed so we know if a new measurement has arrived
-    if (_baro.get_last_update() != lastHgtReceived_ms) {
-        // Don't use Baro height if operating in optical flow mode as we use range finder instead
-        if (frontend._fusionModeGPS == 3 && frontend._altSource == 1) {
-            if ((imuSampleTime_ms - rngValidMeaTime_ms) < 2000) {
-                // adjust range finder measurement to allow for effect of vehicle tilt and height of sensor
-                baroDataNew.hgt = max(rngMea * Tnb_flow.c.z, rngOnGnd);
-                // calculate offset to baro data that enables baro to be used as a backup
-                // filter offset to reduce effect of baro noise and other transient errors on estimate
-                baroHgtOffset = 0.1f * (_baro.get_altitude() + stateStruct.position.z) + 0.9f * baroHgtOffset;
-            } else if (isAiding && takeOffDetected) {
-                // use baro measurement and correct for baro offset - failsafe use only as baro will drift
-                baroDataNew.hgt = max(_baro.get_altitude() - baroHgtOffset, rngOnGnd);
-            } else {
-                // If we are on ground and have no range finder reading, assume the nominal on-ground height
-                baroDataNew.hgt = rngOnGnd;
-                // calculate offset to baro data that enables baro to be used as a backup
-                // filter offset to reduce effect of baro noise and other transient errors on estimate
-                baroHgtOffset = 0.1f * (_baro.get_altitude() + stateStruct.position.z) + 0.9f * baroHgtOffset;
-            }
-        } else {
-            // use baro measurement and correct for baro offset
-            baroDataNew.hgt = _baro.get_altitude();
-        }
-
-        // filtered baro data used to provide a reference for takeoff
-        // it is is reset to last height measurement on disarming in performArmingChecks()
-        if (!getTakeoffExpected()) {
-            const float gndHgtFiltTC = 0.5f;
-            const float dtBaro = frontend.hgtAvg_ms*1.0e-3f;
-            float alpha = constrain_float(dtBaro / (dtBaro+gndHgtFiltTC),0.0f,1.0f);
-            meaHgtAtTakeOff += (baroDataDelayed.hgt-meaHgtAtTakeOff)*alpha;
-        } else if (isAiding && getTakeoffExpected()) {
-            // If we are in takeoff mode, the height measurement is limited to be no less than the measurement at start of takeoff
-            // This prevents negative baro disturbances due to copter downwash corrupting the EKF altitude during initial ascent
-            baroDataNew.hgt = max(baroDataNew.hgt, meaHgtAtTakeOff);
-        }
-
-        // time stamp used to check for new measurement
-        lastHgtReceived_ms = _baro.get_last_update();
-
-        // estimate of time height measurement was taken, allowing for delays
-        hgtMeasTime_ms = lastHgtReceived_ms - frontend._hgtDelay_ms;
-
-        // save baro measurement to buffer to be fused later
-        baroDataNew.time_ms = hgtMeasTime_ms;
-        StoreBaro();
-    }
-}
-
-// store baro in a history array
-void NavEKF2_core::StoreBaro()
-{
-    if (baroStoreIndex >= OBS_BUFFER_LENGTH) {
-        baroStoreIndex = 0;
-    }
-    storedBaro[baroStoreIndex] = baroDataNew;
-    baroStoreIndex += 1;
-}
-
-// return newest un-used baro data that has fallen behind the fusion time horizon
-// if no un-used data is available behind the fusion horizon, return false
-bool NavEKF2_core::RecallBaro()
-{
-    baro_elements dataTemp;
-    baro_elements dataTempZero;
-    dataTempZero.time_ms = 0;
-    uint32_t temp_ms = 0;
-    for (uint8_t i=0; i<OBS_BUFFER_LENGTH; i++) {
-        dataTemp = storedBaro[i];
-        // find a measurement older than the fusion time horizon that we haven't checked before
-        if (dataTemp.time_ms != 0 && dataTemp.time_ms <= imuDataDelayed.time_ms) {
-            // zero the time stamp so we won't use it again
-            storedBaro[i]=dataTempZero;
-            // Find the most recent non-stale measurement that meets the time horizon criteria
-            if (((imuDataDelayed.time_ms - dataTemp.time_ms) < 500) && dataTemp.time_ms > temp_ms) {
-                baroDataDelayed = dataTemp;
-                temp_ms = dataTemp.time_ms;
-            }
-        }
-    }
-    if (temp_ms != 0) {
-        return true;
-    } else {
-        return false;
     }
 }
