+++ libraries/AP_NavEKF2/AP_NavEKF2_Wind.cpp
@@ -22,49 +22,6 @@ extern const AP_HAL::HAL& hal;
 *                   RESET FUNCTIONS                     *
 ********************************************************/
 
-
-
-/********************************************************
-*            GET STATES/PARAMS FUNCTIONS                *
-********************************************************/
-
-// return the NED wind speed estimates in m/s (positive is air moving in the direction of the axis)
-void NavEKF2_core::getWind(Vector3f &wind) const
-{
-    wind.x = stateStruct.wind_vel.x;
-    wind.y = stateStruct.wind_vel.y;
-    wind.z = 0.0f; // currently don't estimate this
-}
-
-/********************************************************
-*            SET STATES/PARAMS FUNCTIONS                *
-********************************************************/
-
-/********************************************************
-*                      READ SENSORS                     *
-********************************************************/
-
-// check for new airspeed data and update stored measurements if available
-void NavEKF2_core::readAirSpdData()
-{
-    // if airspeed reading is valid and is set by the user to be used and has been updated then
-    // we take a new reading, convert from EAS to TAS and set the flag letting other functions
-    // know a new measurement is available
-    const AP_Airspeed *aspeed = _ahrs->get_airspeed();
-    if (aspeed &&
-            aspeed->use() &&
-            aspeed->last_update_ms() != timeTasReceived_ms) {
-        tasDataNew.tas = aspeed->get_airspeed() * aspeed->get_EAS2TAS();
-        timeTasReceived_ms = aspeed->last_update_ms();
-        tasDataNew.time_ms = timeTasReceived_ms - frontend.tasDelay_ms;
-        newDataTas = true;
-        StoreTAS();
-        RecallTAS();
-    } else {
-        newDataTas = false;
-    }
-}
-
 /********************************************************
 *                   FUSE MEASURED_DATA                  *
 ********************************************************/
